Remove Python 2.6 classifier
It’s been ten years since Python 2.7 was released and pygtrie has
not been tested with Python 2.6 for quite a while.  Remove Python
2.6 from the package classifiers list to avoid any confusion in
case the library doesn’t work with that version of the language.
This shouldn’t affect anyone.

While at it, mention pygtrie is Python 2.x and Python 3.x compatible
in the documentation.