Add a 'has_children' parameter in node traversal
Define an additional boolean parameter 'has_children' in _Node.traverse that will help determine whether a node has children or not. Without this, the only way to find whether a node has children was to itertate over the 'children' parameter which is a generator object. For example, consider a StringTrie that holds an entire filesystem within itself and some directories may have millions of files or directories in them. Let's say, during a traversal, for some reason we don't want to traverse a sub-trie of a node, in that case, has_children = bool(children) will not work since it will always return True and iterating over the children generator will multiply the time complexity by O(n). To perform this action within constant time O(1), we need some parameter that tells us if the node has children and introducing 'has_children' solves and we pass 'bool(node.children)' while calling traverse.